+++ R/image.R
@@ -14,22 +14,21 @@ add_border = function(input, pixels = 1, color = 'black', output) {
 #'
 #' Compress PNG/JPEG images with \samp{api.tinify.com}, and download the
 #' compressed images. These functions require R packages \pkg{curl} and
-#' \pkg{jsonlite}. `tinify_dir()` is a wrapper function of `tinify()`
-#' to compress images under a directory.
+#' \pkg{jsonlite}. `tinify_dir()` is a wrapper function of `tinify()` to
+#' compress images under a directory.
 #'
 #' You are recommended to set the API key in \file{.Rprofile} or
 #' \file{.Renviron}. After that, the only required argument of this function is
-#' `input`. If the original images can be overwritten by the compressed
-#' images, you may either use `output = identity`, or set the value of the
-#' `history` argument in \file{.Rprofile} or \file{.Renviron}.
+#' `input`. If the original images can be overwritten by the compressed images,
+#' you may either use `output = identity`, or set the value of the `history`
+#' argument in \file{.Rprofile} or \file{.Renviron}.
 #' @param input A vector of input paths of images.
-#' @param output A vector of output paths or a function that takes `input`
-#'   and returns a vector of output paths (e.g., `output = `[`identity`]
-#'   means `output = input`). By default, if the `history` argument is
-#'   not a provided, `output` is `input` with a suffix `-min`
-#'   (e.g., when `input = 'foo.png'`, `output = 'foo-min.png'`),
-#'   otherwise `output` is the same as `input`, which means the
-#'   original image files will be overwritten.
+#' @param output A vector of output paths or a function that takes `input` and
+#'   returns a vector of output paths (e.g., `output = `[`identity`] means
+#'   `output = input`). By default, if the `history` argument is not a provided,
+#'   `output` is `input` with a suffix `-min` (e.g., when `input = 'foo.png'`,
+#'   `output = 'foo-min.png'`), otherwise `output` is the same as `input`, which
+#'   means the original image files will be overwritten.
 #' @param quiet Whether to suppress detailed information about the compression,
 #'   which is of the form \samp{input.png (10 Kb) ==> output.png (5 Kb, 50\%);
 #'   compression count: 42}. The percentage after `output.png` stands for
@@ -39,17 +38,16 @@ add_border = function(input, pixels = 1, color = 'black', output) {
 #'   compressed before. This argument only makes sense when the `history`
 #'   argument is provided.
 #' @param key The Tinify API key. It can be set via either the global option
-#'   `xfun.tinify.key` (you may set it in \file{~/.Rprofile}) or the
-#'   environment variable `R_XFUN_TINIFY_KEY` (you may set it in
-#'   \file{~/.Renviron}).
+#'   `xfun.tinify.key` or the environment variable `R_XFUN_TINIFY_KEY` (see
+#'   [env_option()]).
 #' @param history Path to a history file to record the MD5 checksum of
 #'   compressed images. If the checksum of an expected output image exists in
-#'   this file and `force = FALSE`, the compression will be skipped. This
-#'   can help you avoid unnecessary API calls.
+#'   this file and `force = FALSE`, the compression will be skipped. This can
+#'   help you avoid unnecessary API calls.
 #' @return The output file paths.
 #' @references Tinify API: <https://tinypng.com/developers>.
-#' @seealso The \pkg{tinieR} package (<https://github.com/jmablog/tinieR/>)
-#'   is a more comprehensive implementation of the Tinify API, whereas
+#' @seealso The \pkg{tinieR} package (<https://github.com/jmablog/tinieR/>) is a
+#'   more comprehensive implementation of the Tinify API, whereas
 #'   `xfun::tinify()` has only implemented the feature of shrinking images.
 #' @export
 #' @examplesIf interactive()
@@ -57,8 +55,8 @@ add_border = function(input, pixels = 1, color = 'black', output) {
 #' xfun::tinify(f)  # remember to set the API key before trying this
 tinify = function(
   input, output, quiet = FALSE, force = FALSE,
-  key = getOption('xfun.tinify.key', Sys.getenv('R_XFUN_TINIFY_KEY')),
-  history = getOption('xfun.tinify.history', Sys.getenv('R_XFUN_TINIFY_HISTORY'))
+  key = env_option('xfun.tinify.key'),
+  history = env_option('xfun.tinify.history')
 ) {
   if (!(is.character(key) && length(key) == 1 && key != '')) stop(
     "The value of the 'key' argument must be a single non-empty character string."